Changes to #logo
There were some instances where the website would fall apart when zooming in/out. The main cause was the logo. Gave logo a width in pixels instead of %. Used margin-top and margin-left to correctly position the logo. Avoided using left:## & top:## becuase it made the logo move out of place when zooming in/out.